+++ lsdpack.cpp
@@ -220,11 +220,11 @@ void record_dump(int argc, char* argv[], bool dmg_mode) {
 }
 
 void print_help_and_exit() {
-    puts("usage: lsdpack [-g] [-r] [-D] [lsdj.gb lsdj2.gb ...]");
+    puts("usage: lsdpack [-g] [-r] [-d] [lsdj.gb lsdj2.gb ...]");
     puts("");
     puts("-g: .gbs conversion");
     puts("-r: raw register dump; optimizations disabled");
-    puts("-D: record using emulated DMG (default: CGB)");
+    puts("-d: record using emulated DMG (default: CGB)");
     exit(1);
 }
 
@@ -235,7 +235,7 @@ int main(int argc, char* argv[]) {
     unsigned flags = 0;
 
     int c;
-    while ((c = getopt(argc, argv, "grD")) != -1) {
+    while ((c = getopt(argc, argv, "grd")) != -1) {
         switch (c) {
             case 'g':
                 puts(".gbs mode enabled");
@@ -245,8 +245,8 @@ int main(int argc, char* argv[]) {
                 puts("register dump mode enabled");
                 dump_mode = true;
                 break;
-            case 'D':
-                puts("DMG mode enabled");
+            case 'd':
+                puts("recording using emulated DMG");
                 dmg_mode = true;
                 break;
             default: